Add conditionals for writerole/clustermonitor
Add `clusterMonitor.create` and `writeRole.create` values
that control the conditional creation of the `clusterMonitor`
deployment and the `spyderbat-cluster-write` clusterrole
respectively. The default of the new values is `true` to not change
the previous behaviour.

The behaviour has changed for the clusterrolebinding, where the
name now includes the release namespace.

These changes are useful for running multiple copies of the chart
in the same cluster, where the first instance creates both the
`clusterMonitor` deployment and the `spyderbat-cluster-write` and
all other copies do not.

This is necessary when the cluster runs on heterogeneous node
with vastly different sizes (e.g. nodes with 8 CPUs along side ones
with 192 CPUs), where the memory and cpu limits configuration need
to be tuned correctly.